+++ src/resources/assets/js/components/Log.vue
@@ -3,39 +3,70 @@
         <div class="modal-dialog modal-lg">
             <div class="modal-content" v-if="selectedTest">
                 <div class="modal-header">
-                    <h3 class="modal-title">{{ selectedTest.name }}</h3>
-                    <button type="button" class="close" data-dismiss="modal" aria-label="Close">
-                        <span aria-hidden="true">
-                            <i class="fa fa-times-circle" aria-hidden="true"></i>
-                        </span>
-                    </button>
+                    <h3 class="modal-title">
+                        {{ selectedProject.name }} - {{ selectedTest.name }}
+                    </h3>
+
+                    <div class="close">
+                        <h3>
+                            <span v-if="selectedTest.state == 'failed' || selectedTest.state == 'ok'" :class="'pull-right badge badge-'+(selectedTest.state == 'failed' ? 'danger' : 'success')">
+                                {{ selectedTest.state }}
+                            </span>
+                        </h3>
+                    </div>
                 </div>
 
                 <div class="modal-body">
-                    <div v-if="selectedTest.log" :class="'btn btn-pill ' + getPillColor('log')" @click="setPanelLog()">
-                        command output
-                    </div>
-                    &nbsp;
-                    <div v-if="selectedTest.run.screenshots"  :class="'btn btn-pill ' + getPillColor('screenshot')" @click="setPanelScreenshot()">
-                        screenshots
-                    </div>
-                    &nbsp;
-                    <div v-if="selectedTest.html" :class="'btn btn-pill ' + getPillColor('html')" @click="setPanelHtml()">
-                        {{ getHtmlPaneName() }}
-                    </div>
-
-                    <div :class="'tab-content modal-scroll' + (selectedPanel == 'log' ? ' terminal' : '')  + (selectedPanel == 'html' ? ' html' : '')">
-                        <div v-if="selectedPanel == 'log'" v-html="selectedTest.log" :class="'tab-pane terminal ' + (selectedPanel == 'log' ? 'active' : '')">
+                    <div class="row">
+                        <div class="col-8">
+                            <div v-if="selectedTest.log" :class="'btn btn-pill ' + getPillColor(constants.LOG_ID_LOG)" @click="setPanelLog()">
+                                command output
+                            </div>
+                            &nbsp;
+                            <div v-if="selectedTest.run.screenshots"  :class="'btn btn-pill ' + getPillColor(constants.LOG_ID_SCREENSHOTS)" @click="setPanelScreenshot()">
+                                screenshots
+                            </div>
+                            &nbsp;
+                            <div v-if="selectedTest.html" :class="'btn btn-pill ' + getPillColor(constants.LOG_ID_HTML)" @click="setPanelHtml()">
+                                {{ getHtmlPaneName() }}
+                            </div>
                         </div>
 
-                        <div v-if="selectedPanel == 'screenshot'" :class="'tab-pane ' + (selectedPanel == 'screenshot' ? 'active' : '')">
+                        <div class="col-4 text-right">
+                            <div @click="runTest(selectedTest.id)" :class="'btn btn-sm btn-'+(selectedTest.state == 'running' ? 'secondary' : (selectedTest.state == 'queued' ? 'warning' : 'danger'))">
+                                <div>
+                                    <span v-if="selectedTest.state == 'running'">
+                                        <i class="fa fa-cog fa-spin fa-fw"></i> running...
+                                    </span>
+
+                                    <span v-if="selectedTest.state == 'queued'">
+                                        <span class="fa fa-clock"></span> in queue
+                                    </span>
+
+                                    <span v-if="selectedTest.state !== 'running' && selectedTest.state !== 'queued'">
+                                        <span class="fa fa-play"></span> run it
+                                    </span>
+                                </div>
+                            </div>
+
+                            <div @click="editFile(selectedTest.edit_file_url)" class="btn btn-sm btn-primary">
+                                open in {{ selectedTest.editor_name }}
+                            </div>
+                        </div>
+                    </div>
+
+                    <div :class="'tab-content modal-scroll' + (selectedPanel == constants.LOG_ID_LOG ? ' terminal' : '')  + (selectedPanel == constants.LOG_ID_HTML ? ' html' : '')">
+                        <div v-if="selectedPanel == constants.LOG_ID_LOG" v-html="selectedTest.log" :class="'tab-pane terminal ' + (selectedPanel == constants.LOG_ID_LOG ? 'active' : '')">
+                        </div>
+
+                        <div v-if="selectedPanel == constants.LOG_ID_SCREENSHOTS" :class="'tab-pane ' + (selectedPanel == constants.LOG_ID_SCREENSHOTS ? 'active' : '')">
                             <div v-for="screenshot in JSON.parse(selectedTest.run.screenshots)" class="text-center">
                                 <h3>{{ String(screenshot).substring(screenshot.lastIndexOf('/') + 1) }}</h3>
                                 <img :src="makeScreenshot(screenshot)" :alt="screenshot" class="screenshot"/>
                             </div>
                         </div>
 
-                        <div v-if="selectedPanel == 'html'"  v-html="selectedTest.html" :class="'tab-pane ' + (selectedPanel == 'html' ? 'active' : '')">
+                        <div v-if="selectedPanel == constants.LOG_ID_HTML"  v-html="selectedTest.html" :class="'tab-pane ' + (selectedPanel == constants.LOG_ID_HTML ? 'active' : '')">
                         </div>
                     </div>
                 </div>
@@ -54,21 +85,26 @@
     import {mapState} from 'vuex';
     import {mapMutations} from 'vuex';
     import {mapActions} from 'vuex';
+    import {mapGetters} from 'vuex';
 
     export default {
-        computed: mapState(['laravel', 'logVisible', 'selectedTest', 'selectedPanel']),
+        computed: {
+            ...mapState(['laravel', 'logVisible', 'constants']),
+
+            ...mapGetters(['selectedPanel', 'selectedTest', 'selectedProject']),
+        },
 
         methods: {
             setPanelLog() {
-                this.$store.commit('setSelectedPanel', 'log');
+                this.$store.commit('setSelectedPanel', this.constants.LOG_ID_LOG);
             },
 
             setPanelScreenshot() {
-                this.$store.commit('setSelectedPanel', 'screenshot');
+                this.$store.commit('setSelectedPanel', this.constants.LOG_ID_SCREENSHOTS);
             },
 
             setPanelHtml() {
-                this.$store.commit('setSelectedPanel', 'html');
+                this.$store.commit('setSelectedPanel', this.constants.LOG_ID_HTML);
             },
 
             getPillColor(button) {
@@ -80,7 +116,7 @@
             },
 
             makeScreenshot(screenshot) {
-                return this.laravel.url_prefix+'/files/'+btoa(screenshot)+'/download';
+                return this.laravel.url_prefix+'/files/'+btoa(screenshot)+'/download?random='+Math.random();
             },
 
             baseName(str) {
@@ -95,9 +131,17 @@
 
             getHtmlPaneName() {
                 return this.selectedTest.html.match(/snapshot/i)
-                    ? 'snapshot'
-                    : 'html';
-            }
+                    ? this.constants.LOG_ID_SNAPSHOT
+                    : this.constants.LOG_ID_HTML;
+            },
+
+            runTest(testId) {
+                axios.get(this.laravel.url_prefix+'/tests/run/'+testId);
+            },
+
+            editFile(file) {
+                axios.get(file);
+            },
         }
     }
 </script>